Revert "Merge remote-tracking branch 'origin/GameOver4' into features-fix"
This reverts commit a5bed78ea5cdeecf88856c04a4cea32b1f382d2f, reversing
changes made to b340dab10468ffab2722a401c314fd218a049024.

+++ Assets/Scripts/MainMenu.cs
@@ -27,12 +27,19 @@ public class MainMenu : MonoBehaviour
     }
 
     public void NuevoJuego(){
-        
-        PlayerPrefs.DeleteAll();
-        
-        player.gameObject.SetActive(true);
 
-        UnityEngine.SceneManagement.SceneManager.LoadScene(NewGameScene);
+        if (PlayerPrefs.HasKey("ContinueLevel")){
+            PlayerPrefs.DeleteAll();
+            UnityEngine.SceneManagement.SceneManager.LoadScene(NewGameScene);
+        }else{
+            Debug.Log("Continuando partida");
+
+            player.gameObject.SetActive(true);
+        
+            player.transform.position = new Vector3(PlayerPrefs.GetFloat("PosX"),PlayerPrefs.GetFloat("PosY"),PlayerPrefs.GetFloat("PosZ"));
+            UnityEngine.SceneManagement.SceneManager.LoadScene(NewGameScene);
+        }
+        
     }
 
     public void Continue(){
